♻️Refactor: 할일 생성, 할일 카드 모달 modal폴더로 이동 및 필요없는 컴포넌트 삭제
