+++ src/components/modal/cardDetail/types/index.ts
